[GEOS-10518] Partial RELINQUISH_LOG4J_CONTROL regression with WildFly
